优化了一些UI，修复了一些bug，增加了更多的bug

+++ H2CO3Resources/src/main/java/org/koishi/launcher/h2co3/resources/component/MenuView.java
@@ -105,7 +105,7 @@ public class MenuView extends View {
         canvas.drawCircle(getMeasuredWidth() >> 1, getMeasuredHeight() >> 1, (getMeasuredWidth() >> 1) - 1, strokePaint);
         canvas.drawCircle(getMeasuredWidth() >> 1, getMeasuredHeight() >> 1, (getMeasuredWidth() >> 1) - 2, areaPaint);
 
-        int iconSize = (int) Math.min(getMeasuredWidth(), getMeasuredHeight() / ICON_SIZE_RATIO);
+        int iconSize = Math.min(getMeasuredWidth(), getMeasuredHeight() / ICON_SIZE_RATIO);
         int iconLeft = (getMeasuredWidth() - iconSize) / 2;
         int iconTop = (getMeasuredHeight() - iconSize) / 2;
         int iconRight = iconLeft + iconSize;
